move lnurl stuff out of component
