+++ src/components/PayInvoiceDialog.vue
@@ -213,7 +213,7 @@ import { useCameraStore } from "src/stores/camera";
 
 import { mapActions, mapState, mapWritableState } from "pinia";
 import ChooseMint from "components/ChooseMint.vue";
-import * as bolt11Decoder from "light-bolt11-decoder";
+// import * as bolt11Decoder from "light-bolt11-decoder";
 import * as _ from "underscore";
 
 export default defineComponent({
@@ -234,7 +234,7 @@ export default defineComponent({
     ...mapState(useWalletStore, ["payInvoiceData"]),
   },
   methods: {
-    ...mapActions(useWalletStore, ["melt"]),
+    ...mapActions(useWalletStore, ["melt", "decodeRequest", "lnurlPaySecond"]),
     ...mapActions(useCameraStore, ["closeCamera", "showCamera", "hasCamera"]),
     canPay: function () {
       if (!this.payInvoiceData.invoice) return false;
@@ -245,156 +245,156 @@ export default defineComponent({
         clearInterval(this.payInvoiceData.paymentChecker);
       }, 10000);
     },
-    decodeRequest: function (r = null) {
-      // set the argument as the data to parse
-      if (typeof r == "string" && r != null) {
-        this.payInvoiceData.data.request = r;
-      }
-      let reqtype = null;
-      let req = null;
-      // get request
-      if (this.camera.data) {
-        // get request from camera
-        req = this.camera.data;
-      } else if (this.payInvoiceData.data.request) {
-        // get request from pay invoice dialog
-        req = this.payInvoiceData.data.request;
-      }
+    // decodeRequest: function (r = null) {
+    //   // set the argument as the data to parse
+    //   if (typeof r == "string" && r != null) {
+    //     this.payInvoiceData.data.request = r;
+    //   }
+    //   let reqtype = null;
+    //   let req = null;
+    //   // get request
+    //   if (this.camera.data) {
+    //     // get request from camera
+    //     req = this.camera.data;
+    //   } else if (this.payInvoiceData.data.request) {
+    //     // get request from pay invoice dialog
+    //     req = this.payInvoiceData.data.request;
+    //   }
 
-      if (req.toLowerCase().startsWith("lnbc")) {
-        this.payInvoiceData.data.request = req;
-        reqtype = "bolt11";
-      } else if (req.toLowerCase().startsWith("lightning:")) {
-        this.payInvoiceData.data.request = req.slice(10);
-        reqtype = "bolt11";
-      } else if (req.toLowerCase().startsWith("lnurl:")) {
-        this.payInvoiceData.data.request = req.slice(6);
-        reqtype = "lnurl";
-      } else if (req.indexOf("lightning=lnurl1") !== -1) {
-        this.payInvoiceData.data.request = req
-          .split("lightning=")[1]
-          .split("&")[0];
-        reqtype = "lnurl";
-      } else if (
-        req.toLowerCase().startsWith("lnurl1") ||
-        req.match(/[\w.+-~_]+@[\w.+-~_]/)
-      ) {
-        this.payInvoiceData.data.request = req;
-        reqtype = "lnurl";
-      } else if (req.indexOf("cashuA")) {
-        // very dirty way of parsing cashu tokens from either a pasted token or a URL like https://host.com?token=eyJwcm
-        this.receiveData.tokensBase64 = req.slice(req.indexOf("cashuA"));
-        reqtype = "cashu";
-      }
+    //   if (req.toLowerCase().startsWith("lnbc")) {
+    //     this.payInvoiceData.data.request = req;
+    //     reqtype = "bolt11";
+    //   } else if (req.toLowerCase().startsWith("lightning:")) {
+    //     this.payInvoiceData.data.request = req.slice(10);
+    //     reqtype = "bolt11";
+    //   } else if (req.toLowerCase().startsWith("lnurl:")) {
+    //     this.payInvoiceData.data.request = req.slice(6);
+    //     reqtype = "lnurl";
+    //   } else if (req.indexOf("lightning=lnurl1") !== -1) {
+    //     this.payInvoiceData.data.request = req
+    //       .split("lightning=")[1]
+    //       .split("&")[0];
+    //     reqtype = "lnurl";
+    //   } else if (
+    //     req.toLowerCase().startsWith("lnurl1") ||
+    //     req.match(/[\w.+-~_]+@[\w.+-~_]/)
+    //   ) {
+    //     this.payInvoiceData.data.request = req;
+    //     reqtype = "lnurl";
+    //   } else if (req.indexOf("cashuA")) {
+    //     // very dirty way of parsing cashu tokens from either a pasted token or a URL like https://host.com?token=eyJwcm
+    //     this.receiveData.tokensBase64 = req.slice(req.indexOf("cashuA"));
+    //     reqtype = "cashu";
+    //   }
 
-      if (reqtype == "bolt11") {
-        console.log("#### QR CODE: BOLT11");
-        this.payInvoiceData.show = true;
-        let invoice;
-        try {
-          invoice = bolt11Decoder.decode(this.payInvoiceData.data.request);
-        } catch (error) {
-          this.notifyWarning("Failed to decode invoice", null, 3000);
-          this.payInvoiceData.show = false;
-          throw error;
-        }
+    //   if (reqtype == "bolt11") {
+    //     console.log("#### QR CODE: BOLT11");
+    //     this.payInvoiceData.show = true;
+    //     let invoice;
+    //     try {
+    //       invoice = bolt11Decoder.decode(this.payInvoiceData.data.request);
+    //     } catch (error) {
+    //       this.notifyWarning("Failed to decode invoice", null, 3000);
+    //       this.payInvoiceData.show = false;
+    //       throw error;
+    //     }
 
-        // invoice.amount = invoice.sections[2] / 1000;
-        // invoice.amount_msat = invoice.sections[2];
-        let cleanInvoice = {};
-        // let cleanInvoice = {
-        //   msat: invoice.amount_msat,
-        //   sat: invoice.amount,
-        //   fsat: invoice.amount,
-        // };
-        // _.each(invoice.sections, (tag) => {
-        //   console.log(tag);
-        // });
-        _.each(invoice.sections, (tag) => {
-          if (_.isObject(tag) && _.has(tag, "name")) {
-            if (tag.name === "amount") {
-              cleanInvoice.msat = tag.value;
-              cleanInvoice.sat = tag.value / 1000;
-              cleanInvoice.fsat = cleanInvoice.sat;
-            } else if (tag.name === "payment_hash") {
-              cleanInvoice.hash = tag.value;
-            } else if (tag.name === "description") {
-              cleanInvoice.description = tag.value;
-            } else if (tag.name === "timestamp") {
-              cleanInvoice.timestamp = tag.value;
-            } else if (tag.name === "expiry") {
-              var expireDate = new Date(
-                (cleanInvoice.timestamp + tag.value) * 1000
-              );
-              cleanInvoice.expireDate = date.formatDate(
-                expireDate,
-                "YYYY-MM-DDTHH:mm:ss.SSSZ"
-              );
-              cleanInvoice.expired = false; // TODO
-            }
-          }
-        });
+    //     // invoice.amount = invoice.sections[2] / 1000;
+    //     // invoice.amount_msat = invoice.sections[2];
+    //     let cleanInvoice = {};
+    //     // let cleanInvoice = {
+    //     //   msat: invoice.amount_msat,
+    //     //   sat: invoice.amount,
+    //     //   fsat: invoice.amount,
+    //     // };
+    //     // _.each(invoice.sections, (tag) => {
+    //     //   console.log(tag);
+    //     // });
+    //     _.each(invoice.sections, (tag) => {
+    //       if (_.isObject(tag) && _.has(tag, "name")) {
+    //         if (tag.name === "amount") {
+    //           cleanInvoice.msat = tag.value;
+    //           cleanInvoice.sat = tag.value / 1000;
+    //           cleanInvoice.fsat = cleanInvoice.sat;
+    //         } else if (tag.name === "payment_hash") {
+    //           cleanInvoice.hash = tag.value;
+    //         } else if (tag.name === "description") {
+    //           cleanInvoice.description = tag.value;
+    //         } else if (tag.name === "timestamp") {
+    //           cleanInvoice.timestamp = tag.value;
+    //         } else if (tag.name === "expiry") {
+    //           var expireDate = new Date(
+    //             (cleanInvoice.timestamp + tag.value) * 1000
+    //           );
+    //           cleanInvoice.expireDate = date.formatDate(
+    //             expireDate,
+    //             "YYYY-MM-DDTHH:mm:ss.SSSZ"
+    //           );
+    //           cleanInvoice.expired = false; // TODO
+    //         }
+    //       }
+    //     });
 
-        this.payInvoiceData.invoice = Object.freeze(cleanInvoice);
-      } else if (reqtype == "lnurl") {
-        console.log("#### QR CODE: LNURL");
-        this.lnurlPayFirst(this.payInvoiceData.data.request);
-      } else if (reqtype == "cashu") {
-        console.log("#### QR CODE: CASHU TOKEN");
-        this.payInvoiceData.show = false;
-        this.showReceiveTokens = true;
-      }
-    },
-    lnurlPayFirst: async function (address) {
-      var host;
-      if (address.split("@").length == 2) {
-        let [user, lnaddresshost] = address.split("@");
-        host = `https://${lnaddresshost}/.well-known/lnurlp/${user}`;
-      } else if (address.toLowerCase().slice(0, 6) === "lnurl1") {
-        let host = Buffer.from(
-          bech32.fromWords(bech32.decode(address, 20000).words)
-        ).toString();
-        var { data } = await axios.get(host);
-        // const { data } = await LNbits.api.request(
-        //   "POST",
-        //   "/api/v1/payments/decode",
-        //   "",
-        //   {
-        //     data: address,
-        //   }
-        // );
-        host = data.domain;
-      }
-      var { data } = await axios.get(host);
-      if (data.tag == "payRequest") {
-        this.payInvoiceData.domain = host.split("https://")[1].split("/")[0];
-        this.payInvoiceData.lnurlpay = data;
-        if (
-          this.payInvoiceData.lnurlpay.maxSendable ==
-          this.payInvoiceData.lnurlpay.minSendable
-        ) {
-          this.payInvoiceData.data.amount =
-            this.payInvoiceData.lnurlpay.maxSendable / 1000;
-        }
-        this.payInvoiceData.show = true;
-      }
-    },
-    lnurlPaySecond: async function () {
-      let amount = this.payInvoiceData.data.amount;
-      if (
-        this.payInvoiceData.lnurlpay.tag == "payRequest" &&
-        this.payInvoiceData.lnurlpay.minSendable <=
-          amount * 1000 <=
-          this.payInvoiceData.lnurlpay.maxSendable
-      ) {
-        var { data } = await axios.get(
-          `${this.payInvoiceData.lnurlpay.callback}?amount=${amount * 1000}`
-        );
-        console.log(data.pr);
-        this.payInvoiceData.data.request = data.pr;
-        this.decodeRequest();
-      }
-    },
+    //     this.payInvoiceData.invoice = Object.freeze(cleanInvoice);
+    //   } else if (reqtype == "lnurl") {
+    //     console.log("#### QR CODE: LNURL");
+    //     this.lnurlPayFirst(this.payInvoiceData.data.request);
+    //   } else if (reqtype == "cashu") {
+    //     console.log("#### QR CODE: CASHU TOKEN");
+    //     this.payInvoiceData.show = false;
+    //     this.showReceiveTokens = true;
+    //   }
+    // },
+    // lnurlPayFirst: async function (address) {
+    //   var host;
+    //   if (address.split("@").length == 2) {
+    //     let [user, lnaddresshost] = address.split("@");
+    //     host = `https://${lnaddresshost}/.well-known/lnurlp/${user}`;
+    //   } else if (address.toLowerCase().slice(0, 6) === "lnurl1") {
+    //     let host = Buffer.from(
+    //       bech32.fromWords(bech32.decode(address, 20000).words)
+    //     ).toString();
+    //     var { data } = await axios.get(host);
+    //     // const { data } = await LNbits.api.request(
+    //     //   "POST",
+    //     //   "/api/v1/payments/decode",
+    //     //   "",
+    //     //   {
+    //     //     data: address,
+    //     //   }
+    //     // );
+    //     host = data.domain;
+    //   }
+    //   var { data } = await axios.get(host);
+    //   if (data.tag == "payRequest") {
+    //     this.payInvoiceData.domain = host.split("https://")[1].split("/")[0];
+    //     this.payInvoiceData.lnurlpay = data;
+    //     if (
+    //       this.payInvoiceData.lnurlpay.maxSendable ==
+    //       this.payInvoiceData.lnurlpay.minSendable
+    //     ) {
+    //       this.payInvoiceData.data.amount =
+    //         this.payInvoiceData.lnurlpay.maxSendable / 1000;
+    //     }
+    //     this.payInvoiceData.show = true;
+    //   }
+    // },
+    // lnurlPaySecond: async function () {
+    //   let amount = this.payInvoiceData.data.amount;
+    //   if (
+    //     this.payInvoiceData.lnurlpay.tag == "payRequest" &&
+    //     this.payInvoiceData.lnurlpay.minSendable <=
+    //       amount * 1000 <=
+    //       this.payInvoiceData.lnurlpay.maxSendable
+    //   ) {
+    //     var { data } = await axios.get(
+    //       `${this.payInvoiceData.lnurlpay.callback}?amount=${amount * 1000}`
+    //     );
+    //     console.log(data.pr);
+    //     this.payInvoiceData.data.request = data.pr;
+    //     this.decodeRequest();
+    //   }
+    // },
   },
   created: function () {},
 });
