+++ src/stores/wallet.ts
@@ -5,6 +5,8 @@ import { useLocalStorage } from "@vueuse/core";
 import { useProofsStore } from "./proofs";
 import { useTokensStore } from "./tokens";
 import { useReceiveTokensStore } from "./receiveTokensStore";
+import { useCameraStore } from "./camera";
+
 import { step1Alice, step3Alice } from "src/js/dhke";
 import * as nobleSecp256k1 from "@noble/secp256k1";
 import { splitAmount } from "src/js/utils";
@@ -13,6 +15,10 @@ import { uint8ToBase64 } from "src/js/base64";
 import token from "src/js/token";
 import { notifyApiError, notifyError, notifySuccess, notify } from "src/js/notify";
 import { CashuMint, CashuWallet } from "@cashu/cashu-ts";
+import * as bolt11Decoder from "light-bolt11-decoder";
+import bech32 from "bech32";
+import axios from "axios";
+import { date } from "quasar";
 
 type Invoice = {
   amount: number;
@@ -704,5 +710,156 @@ export const useWalletStore = defineStore("wallet", {
     //     }
     //   }
     // },
+    decodeRequest: function (r = null) {
+      const camera = useCameraStore();
+      // set the argument as the data to parse
+      if (typeof r == "string" && r != null) {
+        this.payInvoiceData.data.request = r;
+      }
+      let reqtype = null;
+      let req = null;
+      // get request
+      if (camera.camera.data) {
+        // get request from camera
+        req = camera.camera.data;
+      } else if (this.payInvoiceData.data.request) {
+        // get request from pay invoice dialog
+        req = this.payInvoiceData.data.request;
+      }
+
+      if (req.toLowerCase().startsWith("lnbc")) {
+        this.payInvoiceData.data.request = req;
+        reqtype = "bolt11";
+      } else if (req.toLowerCase().startsWith("lightning:")) {
+        this.payInvoiceData.data.request = req.slice(10);
+        reqtype = "bolt11";
+      } else if (req.toLowerCase().startsWith("lnurl:")) {
+        this.payInvoiceData.data.request = req.slice(6);
+        reqtype = "lnurl";
+      } else if (req.indexOf("lightning=lnurl1") !== -1) {
+        this.payInvoiceData.data.request = req
+          .split("lightning=")[1]
+          .split("&")[0];
+        reqtype = "lnurl";
+      } else if (
+        req.toLowerCase().startsWith("lnurl1") ||
+        req.match(/[\w.+-~_]+@[\w.+-~_]/)
+      ) {
+        this.payInvoiceData.data.request = req;
+        reqtype = "lnurl";
+      } else if (req.indexOf("cashuA")) {
+        // very dirty way of parsing cashu tokens from either a pasted token or a URL like https://host.com?token=eyJwcm
+        this.receiveData.tokensBase64 = req.slice(req.indexOf("cashuA"));
+        reqtype = "cashu";
+      }
+
+      if (reqtype == "bolt11") {
+        console.log("#### QR CODE: BOLT11");
+        this.payInvoiceData.show = true;
+        let invoice;
+        try {
+          invoice = bolt11Decoder.decode(this.payInvoiceData.data.request);
+        } catch (error) {
+          this.notifyWarning("Failed to decode invoice", null, 3000);
+          this.payInvoiceData.show = false;
+          throw error;
+        }
+
+        // invoice.amount = invoice.sections[2] / 1000;
+        // invoice.amount_msat = invoice.sections[2];
+        let cleanInvoice = {};
+        // let cleanInvoice = {
+        //   msat: invoice.amount_msat,
+        //   sat: invoice.amount,
+        //   fsat: invoice.amount,
+        // };
+        // _.each(invoice.sections, (tag) => {
+        //   console.log(tag);
+        // });
+        _.each(invoice.sections, (tag) => {
+          if (_.isObject(tag) && _.has(tag, "name")) {
+            if (tag.name === "amount") {
+              cleanInvoice.msat = tag.value;
+              cleanInvoice.sat = tag.value / 1000;
+              cleanInvoice.fsat = cleanInvoice.sat;
+            } else if (tag.name === "payment_hash") {
+              cleanInvoice.hash = tag.value;
+            } else if (tag.name === "description") {
+              cleanInvoice.description = tag.value;
+            } else if (tag.name === "timestamp") {
+              cleanInvoice.timestamp = tag.value;
+            } else if (tag.name === "expiry") {
+              var expireDate = new Date(
+                (cleanInvoice.timestamp + tag.value) * 1000
+              );
+              cleanInvoice.expireDate = date.formatDate(
+                expireDate,
+                "YYYY-MM-DDTHH:mm:ss.SSSZ"
+              );
+              cleanInvoice.expired = false; // TODO
+            }
+          }
+        });
+
+        this.payInvoiceData.invoice = Object.freeze(cleanInvoice);
+      } else if (reqtype == "lnurl") {
+        console.log("#### QR CODE: LNURL");
+        this.lnurlPayFirst(this.payInvoiceData.data.request);
+      } else if (reqtype == "cashu") {
+        console.log("#### QR CODE: CASHU TOKEN");
+        this.payInvoiceData.show = false;
+        this.showReceiveTokens = true;
+      }
+    },
+    lnurlPayFirst: async function (address) {
+      var host;
+      if (address.split("@").length == 2) {
+        let [user, lnaddresshost] = address.split("@");
+        host = `https://${lnaddresshost}/.well-known/lnurlp/${user}`;
+      } else if (address.toLowerCase().slice(0, 6) === "lnurl1") {
+        let host = Buffer.from(
+          bech32.fromWords(bech32.decode(address, 20000).words)
+        ).toString();
+        var { data } = await axios.get(host);
+        // const { data } = await LNbits.api.request(
+        //   "POST",
+        //   "/api/v1/payments/decode",
+        //   "",
+        //   {
+        //     data: address,
+        //   }
+        // );
+        host = data.domain;
+      }
+      var { data } = await axios.get(host);
+      if (data.tag == "payRequest") {
+        this.payInvoiceData.domain = host.split("https://")[1].split("/")[0];
+        this.payInvoiceData.lnurlpay = data;
+        if (
+          this.payInvoiceData.lnurlpay.maxSendable ==
+          this.payInvoiceData.lnurlpay.minSendable
+        ) {
+          this.payInvoiceData.data.amount =
+            this.payInvoiceData.lnurlpay.maxSendable / 1000;
+        }
+        this.payInvoiceData.show = true;
+      }
+    },
+    lnurlPaySecond: async function () {
+      let amount = this.payInvoiceData.data.amount;
+      if (
+        this.payInvoiceData.lnurlpay.tag == "payRequest" &&
+        this.payInvoiceData.lnurlpay.minSendable <=
+        amount * 1000 <=
+        this.payInvoiceData.lnurlpay.maxSendable
+      ) {
+        var { data } = await axios.get(
+          `${this.payInvoiceData.lnurlpay.callback}?amount=${amount * 1000}`
+        );
+        console.log(data.pr);
+        this.payInvoiceData.data.request = data.pr;
+        this.decodeRequest();
+      }
+    },
   },
 });
